requirements.txt
streamlit
google-generativeai
docxtpl
python-docx